+++ src/sphere_simulation.cpp
@@ -10,8 +10,8 @@ sphere_simulation::sphere_simulation(int n) {
   std::random_device rd;
   std::mt19937 gen(rd());
   std::poisson_distribution<int> poisson_dist(n);
-  std::normal_distribution<double> normal_dist(0, 1);
-  std::uniform_real_distribution<double> uniform_dist(0, 1);
+  std::normal_distribution<long double> normal_dist(0, 1);
+  std::uniform_real_distribution<long double> uniform_dist(0, 1);
 
   // Generate a random number from the Poisson distribution
   this->number_of_spheres = poisson_dist(gen);
@@ -20,7 +20,7 @@ sphere_simulation::sphere_simulation(int n) {
   this->torus_size = 1.0;
 
   // data structures
-  this->collision_times = std::vector<double>();
+  this->collision_times = std::vector<long double>();
   this->event_queue = std::priority_queue<Event>();
   this->spheres = new sphere[this->number_of_spheres];
 
@@ -48,7 +48,7 @@ sphere_simulation::sphere_simulation(int n, sphere *spheres) {
   this->current_time = 0.0;
   this->torus_size = 1.0;
   // data structures
-  this->collision_times = std::vector<double>();
+  this->collision_times = std::vector<long double>();
   this->event_queue = std::priority_queue<Event>();
   this->spheres = spheres;
   this->number_of_spheres = n;
@@ -73,7 +73,7 @@ void sphere_simulation::run_simulation() {
   }
   // if there are no more events, update the positions to the max time
   if (current_time < max_time) {
-    double dt = max_time - current_time;
+    long double dt = max_time - current_time;
     update_positions(dt);
     
     for (int i = 0; i < number_of_spheres; i++) {
@@ -100,7 +100,7 @@ void sphere_simulation::handle_event(Event &event) {
   // wrap_around(event.s2);
 
   // Check if spheres still collide at this time
-  double collision_time = collide(event.s1, event.s2);
+  long double collision_time = collide(event.s1, event.s2);
 
   if (collision_time < 0) {
     return; // invalidated by previous collision. discard event.
@@ -123,7 +123,7 @@ void sphere_simulation::handle_event(Event &event) {
   this->find_collision_events(event.s2);
 }
 
-void sphere_simulation::update_positions(double dt) {
+void sphere_simulation::update_positions(long double dt) {
   for (int i = 0; i < number_of_spheres; i++) {
     this->spheres[i].update_position(dt);
     this->wrap_around(&this->spheres[i]);
@@ -131,7 +131,7 @@ void sphere_simulation::update_positions(double dt) {
 }
 
 void sphere_simulation::add_collision_event(sphere *s1, sphere *s2) {
-  double collision_time = collide(s1, s2);
+  long double collision_time = collide(s1, s2);
 
   if (collision_time >= 0) {
     this->event_queue.push(Event(current_time + collision_time, s1, s2));
